This is a test that I can take a picture and display and save the image
This commit allows the user to take a picture and save that image.  Then
once the image is saved it will then be displayed to the user through a
image view.  NOTE:This only works for phones greater than MARSHMELLOW.
I had to create a file path xml to allow the program to access the
storage.  I also had to allow the user to accept or decline permission
to the phone storage.  Some code was commented out in order for the
application to be run.