Update UI and implement library with minimal functionality.
This will update the UI to resemble more of what the app will look like.
It will also still through an arrow after a photo is taken and after you
leave the camera or the gallery.  This has to do with the saving feature
and not being able to restart the imageselector.