Implemented image picker, image model package, and added some functionality to Seefood activity
This commit has a complete implementation of Image model package with
the exception of the convert image to json in image bundle class.  Moved
the image selection to the image bundle classes and added the bitmap to
a arraylist.  This also allows the user to display all the bitmaps in
the arraylist to the application.  I also had moved the display gallery
to its appropriate class. Finally I have removed the Image selection
test classes.