Adjusted image taking and saving features
This commit allows the user to take a picture and save the file path and
file to an image object that is stored in an arraylist of image objects.
Then the same is true for when a user selects an image from the gallery.
I commented out the communication or endpoint class for the mean time
just so this code will compile with my new changes.