Implemented own image picker
This is my version of an image picker.  it allows the user to select one
photo from the gallery and take a picture.  Both of these pictures are
saved as bitmaps as are the file paths. This will also save the image
take in the format that the phone will create such as
month,day,year,hour, and minute.jpg