This solves Rotation bug and imageview not updating correctly
The rotation bug was caused by pictures that were taken were stored with a rotation and I implemented some code to adjust that.  Also when we removed the image from the arraylist it would not go into the if statement under bind images to remove the bitmap from the view.